fix: unable to acquire task when a release happens during its re-acquisition (#831)
# Motivation

Being able to acquire tasks even though it is released while the
re-acquisition.

# Description

When a task is acquired it might be a task from a pod that crashed.
During this acquisition it is possible that the task status will change
as the pod releases the task. This scenario is not being treated by the
actual code.

# Testing

Unit tests yet to be done 

# Impact

It fixes the behaviour of task that are not processed. 

# Checklist

- [x] My code adheres to the coding and style guidelines of the project.
- [x] I have performed a self-review of my code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have made corresponding changes to the documentation.
- [ ] I have thoroughly tested my modifications and added tests when
necessary.
- [x] Tests pass locally and in the CI.
- [ ] I have assessed the performance impact of my modifications.